.circleci/outer-build.sh: update comment; PPC build now uses Rust 1.36
The dep ppv-lite86 started using a construct that was only stabilized in Rust
1.32, and our previous bigendian build image only used 1.31.1.

Note that the name of the downloaded bigendian build image file hasn't
changed, so that this commit doesn't actually change anything in the code; it
might be best to give new build images new filenames. But, I need a commit to
give the CI an excuse to rerun, so here we go.