Cache index and the redirected url
Fixes #85

Motivation
----------

The current approach where tectonic resolves the url and downloads the
full index on every new unavailable file it is requested to fetch leads
to some bad user experience. In particular, compiling a new tex file or
merely adding a figure leads to long compilation delays.

Furthermore, it is a bit scary to use tectonic to batch process a lot of
documents since each of them will lead to a lot of network traffic
(example: I think this is what causes the failure in #430).

We do have --only-cached option but this was introduced as a workaround
for this and it is not desirable to always run tectonic with this
option.

Changes
-------

New CachedITarBundle is a combination of LocalCache and ITarBundle code.
This seems reasonable since only the http ITarBundle backend needs local
caching (at this time).

The main difference is that the the index itself and the resolved
redirected url for the bundle are also cached. Therefore the need to
touch the network is limited only to cases when we are sure that the
file is actually in the tar bundle, and even then we actually only need
to download the requested file (and `SHA256SUM`). This saves the time to
resolve the url (can be slow) and download the index (almost 1MB of
data).

In particular, compiling a new tex document or adding figures will not
have to touch the network at all.

Features
--------

  - Cache the resolved redirected url in `CACHE/redirects/digest_text.txt`.

  - Cache the index in `CACHE/indexes/digest_text.txt`.

  - Touch the online bundle only if the requested file appears in the
    index.

  - Check that the online bundle is what we think it is (checking
    `SHA256SUM`) via the following strategy:

     1. Try to fetch `SHA256SUM` from the cached resolved url based on
        the data in the cached index and compare it to the cached digest.
        If this suceeds, we can be very confident that the tar bundle is
        OK. Advantage: only fetching 64 bytes.
     2. If the above fails, fetch all the data: resolve url, fetch index
        and `SHA256SUM`. Compare with the cached digest. Update the
        resolved url if necessary.
     3. If the above returns a different digest, rewrite the digest and
        exit tectonic (the same behavior to before).

Other changes
------------

- Remove itarbundle.rs and local_cache.rs

- Move cached directory creation into CachedITarBundle::new

- Update cached_itarbundle tests