Add tests for the local cache and tar bundle code
This fills a gap in tectonic's testing coverage by adding tests for the
tar bundle http backend and the local cache. Each test starts a simple
static http server that serves sample tar bundles, for example based on
the content of `tests/assets` directory.

The following tests are implemented:

- `test_full_session`: Runs a full tectonic session. This is a complete
    test but rather slow.
- `test_cached_url_provider`: Tests only the bundle given by
    `make_cached_url_provider`. Fast.
- `test_bundle_update`: Runs a test of the behavior when the tar bundle is
    updated (sha digest changes). Currently an error should be returned.

Since we do not want to pollute the user's cache, we need to point the
local cache root to a temporary directory. This is handled by adding an
optional argument to the `make_cached_url_provider`.

As tectonic does not rely on tar metadata and only requests the file
content from the server, we do not build a real tar bundle to keep the
code simple.