dist: prototype infrastructure to build and deploy the docs/ mdbook
Building is straightforward - there is a binary mdbook artifact we can use.

I've thought about how to deploy and I don't see one obvious best solution. I
can foresee wanting to graft multiple statically-generated HTML trees into the
overall https://tectonic-typesetting.github.io/ URL structure. Here, the model
is that we'll use a repo named `book` that contains only static HTML that is
served at the root https://tectonic-typesetting.github.io/book via GitHub
pages. When books are deployed, the compiled HTML is copied into that
structure (possibly in a subdirectory; e.g. `latest/` and `1.0/`),
amend-committed, and force-pushed up. It's a bit gross but it's the cleanest
way I can think of to get the HTML grafting through GitHub pages. And of
course all of the important history is stored in this, the source repository.