Do not compute the sha256 digest for files assumed read-only
Adds `InputHandle::read_only: bool` field and `fn
InputHandle::new_read_only` that sets this field to `true`. It indicates
that it is not possible to write into such file since the `IoProvider` is
read-only, for example, the format file, the tar bundle or zip bundle,
and so `InputHandle` does not need to compute sha256 digest.

Note that if using a stack of `IoProvider`s, it is still possible to
write into the file if one of the earlier `IoProvider`s allows writing.
In such a case, tectonic will assume that the content of the file
changed since the read digest is unavailable. But it seems very unlikely
that a tex file would be such that it would create a file with the same
name and content as one of the files in the bundle.

Saves 15--25% of tectonics runtime for a typical document.