driver.rs: Report only tex warnings during the last pass
Before this commit, the warnings that happened during an earlier pass
were reported. However, the first pass is expected to have warnings like
undefined references and therefore this is not very useful. After this
commit, tex warnings or ignored errors are reported only if the last tex
pass had warnings.