cfg_support: rebuild this with a fully flexible macro
I got a little obsessed with the challenge of developing a macro that could
support fully general `cfg!` syntax. I came up with something, but the
solution involves basically restructuring the entire module. In the end,
though, we have a single static initialization of the environment variables
and don't need to build complex structures representing each test, so it feels
like a nice solution.