dist/travis.sh: fix release-tag detection logic
It turns out that for tag builds, $TRAVIS_BRANCH is set to the tag name, not
(e.g.) `master`. Which actually makes sense since a tag need not be associated
with any particular branch. So, remove that bit of the conditional.

(cherry picked from commit 376eabc6672b57810744ebe622f632fd5d375227)