Specify target directory implied by the `package` function
By default, cargo places build artifacts in a `target` directory
in the root of the current workspace. This behaviour can be
overridden through various means like ENV variables.

If that is the case, installation of the package will fail because
the tectonic binary will not be in the path assumed by the `package`
function.

We fix this by specifying the path of the `target` directory directly
for the `build` and `check` functions.