Add warning api
The idea is that a Diagnostic is essentially a "builder" for a
warning/error - we apply changes to it (e.g. adding strings) and
eventually output it with diag_finish.

Doing it this way allows us to add more information to warnings later,
such as the filename & line number. Having individual diagnostic_t
objects instead of a global warning allows multiple warnings to be
constructed simultaneously (e.g. across multiple functions).