dist: rework deployment stage to have many jobs
It turns out that we don't need to limit ourselves to a single special
`deployment` job. I don't want to spend forever futzing with this stuff,
but this change should be very helpful because it will make the pipeline
more robust to transient failures: not only can we avoid a failure in
one operation just because it was serialized after another; we can even
potentially retry failed operations through the pipelines UI, if the
issue really was an external transient one.