Update xdvipdfmx to the TeXLive 2020.0 reference
There is a lot going on here so I hope that I got it all right. There is
new code relating to CMaps that alters the PDF output, so unfortunately
it does not seem practical to try to make it so that the new code can
pass the existing PDF-output test suite.