Implement the new \filemoddate primitive
This requires us to add a new feature to the I/O API and so involves
touching a lot of code. In particular, I think that it is probably
important for the memory I/O layer to produce meaningful-ish mtimes,
which adds some complexity there.

The `InputFeatures` trait has added a new method, `get_unix_mtime`, to
support this primitive. It has a default implementation to return
`Ok(None)`, which indicates that the input does not have a well-defined
modification time.

**BREAKING CHANGE**

In order to implement file modtimes for the memory I/O provider, we need
to add more data to its data structures, breaking the APIs for users of
its `files` field. This is a very clear signal that this API should be
cleaned up and made more future-prof. That being said, migration should
be pretty easy: instead of `files` containing a bunch of `Vec<u8>`s, it
now contains a bunch of `MemoryFileInfo` structs that contain a
`Vec<u8>` field named `data`. So you just need to add some `.data` field
accessors to existing code.