src/engines: incorporate remaining data into digest when closing input files
It seems that in some esoteric cases, there are TeX files that are
read-then-written from one run to another, but they are only partially
read. Because our rerun-detection logic compares what was read to what
was written, this sort of incomplete reading triggers a rerun loop.

Here we attempt to fix that by scanning through any remaining unread
data when closing an input file and updating the digest if needed. This
would also yield the right results for the common TeX pattern of opening
an input file and immediately closing it, which we had to special-case
before.

Closes #679.