crates/errors: add a low-level error-type crate
As we break the implementation into smaller crates, the current
error-chain error type used by the main crate becomes untenable, because
it is an enum that explicitly publicizes all of the sub-types that may
contribute to it. That prevents us from building up functionality in
tightly-focused crates, and it prevents us from expressing optional
dependencies well.

I've been very happy with my Cranko error implementation, so here I rip
that off. I'd also be interested to add a feature along the lines of
that provided by the `context-attribute` crate, but that's a bit more
change than I'd like to take on just now.