crates/status_base: crate for basic status-reporting types
Now that we have our basic boxed error type, we can implement our
basic status-message scheme.

As noted in the crate docs, this is *really* similar to reinventing the
wheel of the many logging crates that are out there ... but if you want
to provide nice colorized support in the CLI and other bits of UX
polish, you need something more sophisticated than general logging
tools.