crates/io_base: crate for basic I/O types
Now that we have errors and status reporting, we can separate out the
base I/O system.

We're starting to get non-trivial amounts of code here. In particular,
this crate provides implementations for `flate2` and so must depend on
it. This is because the implementation of a trait for a type most occur
in either the crate that provides the trait, or the crate that provides
the type. flate2 isn't about to start caring about Tectonic's types so
we need to provide the implementation here. This should probably become
an optional feature, but I'm trying to minimize API changes as we move
code around.