BREAKING CHANGE: migrate main crate to use new support crates
Here we take the code that was split into support crates and migrate
the main crate to start using it.

We break a lot of API here, mainly because we have to start using new
error types. This in turn is because we can't put our error_chain error
type low in the stack since it explicitly enumerates all of the
potential library error types that might feed into it. That's a
non-starter in a multi-crate model.

Another breaking change is removing `NoopStatusBackend::new()`, which
should have been `::default()` all along.