tectonic: audit use of ICU headers a bit
Getting a handle on where exactly we use ICU. There are a few uses of
`ucnv` for character-set conversion, and then xetex-ext.c uses ICU For
Real for bidi and linebreaking. We could probably replace ucnv with pure
Rust, but definitely don't want to do that with the other uses.