BREAKING CHANGE: stop using OsStr(ing) for internal filenames
I started out using OsStr/OsString for filenames handled by the TeX
engines, but since XeTeX is Unicode-based there's no way for it to
generate filenames that can't be expressed as UTF-8 (at least, as far as
unusual encodings are concerned, not unrepresentable Unicode
codepoints). The OsStrs are a pain to deal with so I've wanted to fix
this for a while.

Since we now have the `tectonic_io_base` crate that exposes these APIs,
now is the time to do it, before they get further enmeshed in more
public interfaces.