crates/engine_bibtex: import BibTeX as a standalone crate
Now that we have the core FFI bridge stuff modularized, let's see if it
works! Here we pick off the BibTeX functionality and build it as a
standalone crate.

Note that this still isn't wired up in the main Tectonic build as of
yet. Integrating that will be fun since we'll have to port over all of
the C/C++ code in a big batch.