Fix the global setjmp/longjmp, hopefully
Argh, I forgot that you can't just return the result of `setjmp()` --
you have to branch on it at the point of call. I think. Try an
alternative approach for our engine-globals API. I'm not confident in
the semantics of passing around a pointer to a jmp_buf, but it seems to
work on my machine ...