crates/xetex_layout: add this
This breaks the majority of the XeTeX font loading and layout code out
into its own crate, since it is very nearly orthogonal to the rest of
the main XeTeX engine. This opens up the vision of eventually migrating
all of this code to be Rust-based instead of C++-based.

To actually integrate this into the main `tectonic` crate will take some
rearrangement of headers and whatnot. I also commented out one stanza of
tracing code that broke the separation between this code and the core
TeX engine.