BREAKING: fix up include-dir handling
It turns out that in order to get the C/C++ crate-bridging to work
reliably with headers, we can't get away with exposing just a single
include directory from one crate to another; we have to expose a list of
include directories. So, change all of the relevant cases to be
semicolon-separated lists instead of single directory names. Not my
favorite but I don't see a way around it.