Tectonic.toml: preamble, index and postamble
This enables the use of preamble, index and postamble keys in an output.
These get built together to form the input buffer and default to the
current defaults too.

There isn't any fancy checking here as we defer it to the latex engine.
In theory this could enable injection but I'm not aware that would be a
concern at the moment.