io_base: treat filesystem directories as "not found"
This addresses #754, a hard-to-interpret failure and error message when
the engine tries to treat a directory named "latex" as the format file.
Here we change the filesystem code to simply ignore directories. I think
this is the best solution, since it means that we actually just run
correctly when such a directory exists, as opposed to failing but with a
more helpful error message.