crates/pdf_io: move the vast majority of xdvipdfmx into here
This lays the groundwork for us to move the rest of the C engine code,
namely XeTeX and xdvipdfmx, into separate crates. The key thing is that
the way we have set things up, XeTeX needs a bunch of the xdvipdfmx
code, because it uses that to load up PDF files as images, which is a
super important functionality.

I was hoping that at least the XDV portion of xdvipdfmx could be kept
isolated, but it turns out that the PDF-reading code cross-links with
the XDV code via the `dvi_is_tracking_boxes()` function. It wouldn't be
too hard to rework this API to break the interdependency, but I'd rather
make as few modifications as I can get away with for the time being, so
I haven't done that.