bridge_core(!): replace IoEventBackend with DriverHooks
BREAKING: replace the IoEventBackend trait with a DriverHooks trait that
is similar but has some important differences. In particular,
DriverHooks is a single reference that provides combined access to the
functionality that was previously split between the IoProvider and the
IoEventBackend. This unification allows operations that previously
wouldn't have been possible because they'd have required simultaneous
access to both of these.

Philosophically, the DriverHooks trait better captures its purpose:
gathering all of the callbacks needed to help the C/C++ engines
interface with the outside world. IoProvider functionality is *most*,
but not all of this.

Furthermore, in the DriverHooks paradigm, a lot of the events in
IoEventBackend become superfluous. The only ones that can't be captured
with a custom IoProvider implementation are the notifications for when
input or output handles are closed.