tectonic(!): port to use DriverHooks
Port the main crate to use the new DriverHooks interface. The main
difference here is that IoSetup and IoEvents are merged into a single
BridgeState struct, which becomes a "wholly owned subsidiary" of the
ProcessingSession type.

The core insight here is that what we need to do is collect all of the
session functionality that has to be accessed by the C/C++ engines --
the DriverHooks implementation, essentially -- in one struct, which is a
subset of the overall ProcessingSession state. When a C/C++ engine is
running, it maintains a mutable reference to that *subset* of state, but
it makes our life a lot easier if we can isolate that bit of the state
from the overall ProcessingSession struct. Now, this subset must include
the ProcessingSession's whole I/O stack, so it's a non-trivial subset,
but overall the boundaries of the DriverHooks trait help us isolate what
has to go in that subset and what can stay separate.