bridge_core(!): add a shell-escape driver hook
BREAKING: Here we change shell-escape to be implemented by invoking a
new driver hook. This addition on its own need not be a breaking change,
and indeed the DriverHooks trait is designed to allow this to be the
case. However, since we've already accumulated some breaks, we go ahead
and breaks the related C/C++ API and also the unrelated API of the
file-closed event APIs, which really ought to take the status backend as
a parameter in the same way that virtually all of the other I/O APIs do.