engine_xetex(~): track shell-escape API updates
This goes back to the older boolean shell-escape API, which is all we
need with the iterated architecture. We also track the C/C++ API change
that was just made in bridge_core. Also we turn off diagnostic printouts
when shell-escape is enabled, since the Tectonic driver can give much
nicer diagnostics about what's happening there.

This would be a breaking change, but the previous API was never
released.