engine_xetex: fix #714
Due to a mis-transcription, \XeTeXfeaturename and \XeTeXselectorname
were not properly implemented. Thanks @burrbull!

Furthermore, even if they were properly implemented, they would not
have executed currectly due to what appears to be a bug in upstream
XeTeX relating to magic constants getting out of sync:

https://sourceforge.net/p/xetex/bugs/174/

Fixing this bug requires a bump in the format file serial.