bridge_core: expose filesystem paths for just-opened inputs
Leveraging the new "abspath" support in io_base, now track the absolute
path of the most recently-opened input and expose it to the C code with
the new function `ttstub_get_last_input_abspath()`. The plumbing is a
bit messy but fundamentally we're just passing an additional piece of
information around.