tectonic: bring in the SyncTeX fixes
First we need to update our IoProvider implementations to properly
handle the new "abspath" APIs -- providers that delegate to other
providers have to make sure to not just use the default implementation
instead.

We also need to update our test framework to handle the fact that the
contents of the emitted SyncTeX files will depend on the system running
the test suite.