src/driver.rs: add a disable_insecure() method
The idea is to give us a one-stop shop for disabling all known-insecure
engine features, such as the recently-added "shell escape"
functionality. It should be used unless the input data are completely
trusted. It is *not* always turned on, however, because some of these
insecure features deliver functionalities that users like.