tectonic(!): start using the "document model"
We have now split the handling of `Tectonic.toml` files into a separate
crate, `tectonic_docmodel`, so that other people can manipulate
documents without needing to link to the main crate and all of the extra
dependencies that it entails. This ends up working out pretty nicely, in
that we can maintain functionlity with a few targeted extension traits
that wire up the document settings to the actual engine invocations.

However! This is a BREAKING CHANGE because the `tectonic::document` and
`tectonic::workspace` modules have been removed. We don't need or want
them anymore! I doubt anyone has been using them, but if so, use
`tectonic_docmodel` instead, with the new extension traits in the
`tectonic::docmodel` module if needed.