tectonic: fix build for no-serialization-feature
Fine, fix things up so that we build when the `serialization` Cargo
feature is disabled. This is getting annoying to support and I'm not
sure if it's useful anymore (we can build statically linked with proc
macros by adopting a cross-compilation model), but it wasn't too hard to
fix up here.