bridge_core: add a basic security infrastructure
After adding shell-escape I started feeling like we should have a bit
more of a systematic approach to ensuring that API users do their due
diligence with regards to security. Here, we minorly break the crate
behavior to default to disabling all known-insecure features (currently,
only shell-escape), but give a new mechanism to override that. The setup
is centralized such that whenever you try to create a security policy,
it can be overridden in a more conservative direction by setting the
environment variable `TECTONIC_UNTRUSTED_MODE`.