tectonic: use the new security API
Building off of the new core bridge API, we can provide a bit of a more
refined approach to security controls. `tectonic -X compile` and
`tectonic -X build` now take an `--untrusted` option that ensures that
`-Z shell-escape`, and any future insecure features, cannot be enabled.

As with the core bridge work, this changes the ProcessingSessionBuilder
to be in untrusted mode by default, and adds a new API to configure in a
more-trusted mode if that's what you want to do.