Ignore all errors when removing cbindgen Cargo.lock in build.rs
To be able to publish our crates individually, we need to have our build
scripts manually remove a Cargo.lock file that may be created by
cbindgen. I tried to report errors besides ENOENT, but on docs.rs it
seems that the error is EROFS, and I don't want to jump through the
hoops to avoid reporting that one too. So just ignore all errors on the
removal operation.