Switch to dev-time rather than build-time cbindgen
Following up on #788, my no-network workaround seemed to work OK, but
then I ran into problems with `cargo metadata` wanting to create a
Cargo.lock file in a read-only filesystem. I don't see any way to
prevent the cbindgen build.rs invocation from triggering this.

But ... we can just use cbindgen as a standalone tool and save the
generated headers in the repo. I liked the idea of doing things
on-the-fly in build.rs, but I can't think of any reason that we *have*
to, and the dev-time approach is faster and simpler, and it turns out
that it also solves out docs.rs problems. So it's a pretty clear
decision.

Also, I don't expect these generated headers to vary often. But I
documented how to regenerate them in their respective crate READMEs.