tectonic: add support for biber
Using the same kind of approach as used for shell-escape. We start a bit
of a framework for adding passes implemented with external tools. I have
a thought that this can be come a pluggable framework useful in
`Tectonic.toml`, but I need to think a bit about use cases. Also, we'd
need to add some more security handling to restrict whether documents
can cause Tectonic to run arbitrary commands (although it would be more
controllable than shell-escape, at least).