Improve biber handling by parsing the `.run.xml` file
It turns out that in some cases, biber needs access to various `.bib`
files that may be associated with the document build. Fortunately, the
`.run.xml` file contains exactly the information we need to populate the
temporary tool directly properly. We have to link in an XML parser, but
so be it.

I wasn't aware of the `.run.xml` file before and we potentially could
use it a lot more aggressively. We don't do that here, but the support
is written to try to be future-conscious.