bridge_graphite2: fix builds on Rust >=1.55
It seems that as of this release we must use the environment variable
CARGO_ENCODED_RUSTFLAGS instead of just RUSTFLAGS if we want to probe
those. And we do, to determine whether GRAPHITE2_STATIC should be
defined. Fall back to the old variable so that we can still build
with older versions.