geturl: update redirection logic to unbreak archive.org resolution
The archive.org PURL service made a recent change that broke our URL
resolution algorithm. The new redirect structure is:

```
https://archive.org/services/purl/net/pkgwpub/tectonic-default -> 307 Temporary redirect
https://purl.prod.archive.org/net/pkgwpub/tectonic-default -> 302 Found
https://ttassets.z13.web.core.windows.net/tlextras-2020.0r0.tar -> 200 OK
```

The problem is that we stop chasing redirects if the final path
component of the new URL doesn't contain a `.`, in an attempt to not
scan into things like S3 storage URLs. But this means that we stop at
the first URL, which is wrong. To fix this, we semi-hack the logic to
also follow the redirection if the filename of the new URL matches the
filename of the original URL.