geturl: accept a final HTTP status code of 307
This was surfaced by the recent archive.org issue, but doesn't actually
have anything to do with the proper solution of the issue. But while
we're at it, the 302 and 307 status codes are nearly semantically
equivalent, so if we accept 302 we should accept 307. Note that this fix
would have masked the archive.org issue ... things would have worked but
our URL resolution would have been giving improper results.