engine_xetex: start using header autogenerated by the xetex_format crate
We commit the header to Git since it will be useful to look at in the
source tree. The main advantage here is maintainability, and also the
primitive initialization becomes table-driven. Making that change leads
us to reorder the ways in which primitives are set up, which alters the
string table in the format file, and so alters the SHA256 of the format
file without altering its overall semantics.