xetex_format: completely rework the internals
Our first pass was OK, but fell down badly when considering that various
enumerations will evolve between engine versions as subcommands and
parameters are added. It also provided no real support for introspecting
TeX commands -- if the crate doesn't know the actual numerical values of
the various enumerations and command sub-codes, it can't parse them
properly.

This update to the crate adds a massive amount of new infrastructure to
make sure that it can actually introspect commands at runtime. We add a
"symbol table" concept and gradually build up numeric values for
everything we need. We also implement a type for every main TeX command
so that we can actually introspect with flexibility.