engine_xetex: use the new xetex_format
Here we update the generated headers to derive from the latest
xetex_format. A bunch of constants that used to be hardcoded are now
properly computed on-the-fly. We have to bump the format version because
I had a slightly wrong numerical code for \XeTeXvariationname, and the
digest of the plain format file changes because of both that and the
fact that the string table gets reordered significantly with a new
ordering for the primitives initializer table.