Remove time_t from FFI APIs
As per rust-lang/libc#1848, we can't safely do this since the definition
of time_t depends on which version of Musl you're using on 32-bit
systems (the "time64" update for Musl 1.2), and this actually started
biting us on the arm-unknown-linux-musleabihf cross build.
Unfortunately, ISO C makes virtually no guarantees about what time_t is,
so I don't see a practical way to handle the different possibilities
robustly. We'll just cast from i64 or u64 to time_t in C and hope for
the best.