dist: move to three-stage Azure Pipelines build
Here we update to what's now my standard for Cranko-powered builds: a
three-stage process, with Prep, Build&Test, and Deploy. The proximate
motivation is that it looks like maybe the Windows builds are erroring
out because Cranko is detecting Git submodules as dirty files, but it's
a better system anyway. In particular, if/when we start registering on
Zenodo and getting DOIs, we need a single prep stage so that all of the
builds can see and use the same DOI.