Default Windows vcpkg to using a custom triplet
This is basically a tactic to speed up the Windows builds. By default, vcpkg
builds both debug and release versions of requested libraries. There's probably
some good reason for that, but for us it just makes the builds a lot slower than
they need to be. We can avoid that by providing a custom "triplet" file that
disables the debug builds. In order to do this cleanly, we use new support from
cargo-vcpkg that adds support for an "overlay triplets" directory. Builds also
need to set the $VCPKGRS_TRIPLET environment variable so that vcpkg-rs can find
the packages it needs. We add some docs and a diagnostic warning if that may be
what's going on when a vcpkg fails to be found.