engine_spx2html: refactor font-family handling
Transport the code that deals with a document's ensemble of font
families into its own file. There is a lot of code churn here, but I've
compared the before-and-after HTML output with the tt-weave documents
weave.web and xetex.web, and there are no changes besides some places
where `/` and `'` become `&#x2F;` and `&#x27;`.

This is motivated by my plan to serialize the "assets" generated/used by
a spx2html runs, but it also helps with some of the algorithms. By
separating the "fonts" and "content" into separate substructures of the
state, we can tidy up some of the algorithms with regards to
simultaneous mutability.