engine_spx2html: in precomputed-assets mode, ensure that all fonts get loaded
We need to do this to ensure that we can output all of the font assets
at the end of the run, if needed. This depends on a bunch of the
internal reorganizations that I've been doing with font loading. It
still feels pretty messy to me, but I think it's now working.