bridge_harfbuzz: no longer require ICU
The motivation for this change is to fix the Windows pkg-config CI,
where MSYS2's switch to the `pkgconf` implementation of `pkg-config` has
surfaced challenges in this crate's build script. But, it is also true
that we only need harfbuzz-icu for versions of Harfbuzz that are very
old at this point, and separating out that dependency will help us work
towards potentially one day being able to swap out the ICU dependency
with something else. So it's good to make that jump for bigger reasons
than just the CI.