Prefer `tectonic-biber` as the biber executable
When an executable called `tectonic-biber` is found, either in the
current working directory or in `$PATH`, prefer that as the biber
executable over the plain old `biber`. This may help resolve possible
version mismatch between `biber` and the bundled `biblatex` files when
there are multiple TeX installations on a system.