Fix incomplete attached file problem
Currently, we only have `ttstub_input_read` function, which corresponds to
Rust's `read_exact`. This seems to work for most of xetex, but it's not
semantically a replacement for buffered reading with `fread`, which is
used in a dvipdfmx.

One of such places is reading a file to be included in a PDF file as
attachment. This is done through dvipdfmx special like this:

    \special{pdf:fstream @SourceFile (example.txt)}

These reads are done in 1024 byte chunks, and if we read these chunks
with read_exact, we are going to fail unless the file size is not
exactly divisible by 1024.

Tectonic would report this as:

    warning: 1024-byte read failed
    caused by: failed to fill whole buffer

To fix this issue in a non-intrusive way, we introduce a new variant of
reads - `ttstub_input_read_partial`, which follows Rust's `read`
semantics, and not `read_exact`'s, and we use it for the dvipdfmx
fstream reading.

Likely any place that calls `ttstub_input_read` in a loop is also a
candidate for using `ttstub_input_read_partial`, and perhaps it would
work better for others.

But there definitely are places in xetex that depend on the `read_exact`
semantics. For example, when I naively just tried to make all reads into
`ttstub_input_read_partial`, xetex failed pretty quickly on "undumping
the format file", which has a few initial 4 byte reads, but then it
tries to undump big string pool which was many kilobytes in size. Rust
`read` however would only return 8196 minus couple of the the 4 bytes,
which would fail a check on the TeX side which expects full read. AFAICT
xetex uses `fread`, so this probably just comes down to buffering
differences between `fread` and Rust's `Reader` / `BufReader`.

Relates to #935
Fixes #1260