Fixed a regression brought in #1123
When introducing the `inputs` config key, the default behavior when none
of the `inputs`, `preamble`, `index` and `postamble` key is given has
been changed to use an empty list instead of the `DEFAULT_INPUTS` new
default `const`.

Rolled back the default `DEFAULT_INDEX_FILE`, `DEFAULT_PREAMBLE_FILE`
and `DEFAULT_POSTAMBLE` variables to be used when no `inputs` is given.

Signed-off-by: Oror <Lugrim@users.noreply.github.com>